Revert "Frontend branch"
