+++ frontend/src/App.js
@@ -1,15 +1,12 @@
-import logo from './logo.svg';
-import FlowWithProvider from './components/FlowWithProvider';
-import 'reactflow/dist/style.css';
 import './App.css';
+import logo from './logo.svg';
+import Flow from './Flow';
+
 
 function App() {
-
   return (
     <div style={{height:"100vh"}}>
-      <div style={{height: '100%'}}>
-        <FlowWithProvider/>
-      </div>
+      <Flow/>
     </div>
   );
 }
