--- frontend/src/components/customNodes/GoalNode.js
@@ -1,26 +1,0 @@
-import React, { memo } from 'react';
-import { Handle, Position } from 'reactflow';
-
-export default memo(({data, isConnectable, selected}) => {
-    let nodeClassname = 'goalNode';
-    if (selected) nodeClassname += ' selected';
-    return (
-        <>
-            <Handle
-            type="source"
-            position={Position.Bottom}
-            isConnectable={isConnectable}
-            />
-
-            <div className={nodeClassname} onClick={() => console.log(selected)}>
-                { data.label }
-            </div>
-
-            <Handle
-            type="target"
-            position={Position.Top}
-            isConnectable={isConnectable}
-            />
-        </>
-    );
-});
